+++ commands/pets.js
@@ -1,4 +1,4 @@
-const {MessageEmbed} = require('discord.js');
+const { MessageEmbed } = require('discord.js');
 const locale = require('../locale/text.js');
 const petinfo = require('../pets/pets.json');