+++ commands/use.js
@@ -1,5 +1,6 @@
 const locale = require('../locale/text.js');
 const Item = require('../items/item.js');
+const {tradingUsers} = require('../commands/trade.js');
 
 module.exports = {
     name:'use',
@@ -7,6 +8,7 @@ module.exports = {
     syntax:'/use [item name]',
     admin:false,
     async execute(message,args,db){
+        if(tradingUsers[message.author.id]) return message.channel.send(locale.text({lang:db.lang,msg:"in_trade"}));
         if(!args[0]) return message.channel.send(locale.text({lang:db.lang,msg:"syntax"})+this.syntax);
         let userpets = await db.petsdb.get(message.author.id), item = Item.searchItemByName({args:args});
         if(!item) return message.channel.send(locale.text({lang:db.lang,msg:"item_not_found"}));